chore(franchise): separate franchiseSchedule methods into a separated service
